ルート定義を機能ごとのファイルへ分割
app_routes.dart に集約していた RouteData を、タブ（機能）ごとのファイルへ
移動する。app_routes.dart にはパス設計とルートツリー、シェル/ブランチの
定義だけを残し、各画面の構築は機能ごとのファイルが持つようにする。

講義タブと科目検索タブで重複していた科目詳細の構築処理は、過去問を開く
ルートを引数で受け取る buildSubjectDetail として共通化する。
