+++ apps/dotto/lib/router/routes/app_routes.dart
@@ -1,29 +1,11 @@
-import 'dart:async';
-
 import 'package:dotto/domain/tab_item.dart';
-import 'package:dotto/feature/announcement/announcement_detail_screen.dart';
-import 'package:dotto/feature/announcement/announcement_screen.dart';
-import 'package:dotto/feature/bus/bus_screen.dart';
-import 'package:dotto/feature/bus/bus_stop_select.dart';
-import 'package:dotto/feature/bus/bus_trip_screen.dart';
-import 'package:dotto/feature/course/course_cancellation_screen.dart';
-import 'package:dotto/feature/course/course_customize_screen.dart';
-import 'package:dotto/feature/course/course_document_screen.dart';
-import 'package:dotto/feature/course/course_registration_screen.dart';
-import 'package:dotto/feature/course/course_screen.dart';
-import 'package:dotto/feature/debug/debug_screen.dart';
-import 'package:dotto/feature/funch/funch.dart';
-import 'package:dotto/feature/github_contributor/github_contributor_screen.dart';
-import 'package:dotto/feature/map/map_screen.dart';
-import 'package:dotto/feature/onboarding/onboarding_screen.dart';
 import 'package:dotto/feature/root/root_screen.dart';
-import 'package:dotto/feature/setting/settings.dart';
-import 'package:dotto/feature/setting/widget/license.dart';
-import 'package:dotto/feature/subject/search_subject_screen.dart';
-import 'package:dotto/feature/subject/subject_detail_past_exam_screen.dart';
-import 'package:dotto/feature/subject/subject_detail_screen.dart';
-import 'package:dotto/feature/subject/subject_review_new_screen.dart';
-import 'package:dotto/widget/cloudflare_pdf_viewer.dart';
+import 'package:dotto/router/routes/bus_routes.dart';
+import 'package:dotto/router/routes/course_routes.dart';
+import 'package:dotto/router/routes/funch_routes.dart';
+import 'package:dotto/router/routes/map_routes.dart';
+import 'package:dotto/router/routes/setting_routes.dart';
+import 'package:dotto/router/routes/subject_routes.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
@@ -300,490 +282,7 @@ final class SubjectShellBranchData extends StatefulShellBranchData {
   const SubjectShellBranchData();
 }
 
-final class CourseRouteData extends GoRouteData with $CourseRouteData {
-  const CourseRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const CourseScreen();
-  }
-}
-
-final class CourseCustomizeRouteData extends GoRouteData
-    with $CourseCustomizeRouteData {
-  const CourseCustomizeRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const CourseCustomizeScreen();
-  }
-}
-
-final class CourseRegistrationRouteData extends GoRouteData
-    with $CourseRegistrationRouteData {
-  const CourseRegistrationRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const CourseRegistrationScreen();
-  }
-}
-
-/// 休講・補講・教室変更のパスは休講のタブへ寄せる。
-final class CourseNoticeRouteData extends GoRouteData
-    with $CourseNoticeRouteData {
-  const CourseNoticeRouteData();
-
-  @override
-  FutureOr<String?> redirect(BuildContext context, GoRouterState state) {
-    return const CourseNoticeCancellationsRouteData().location;
-  }
-}
-
-final class CourseNoticeCancellationsRouteData extends GoRouteData
-    with $CourseNoticeCancellationsRouteData {
-  const CourseNoticeCancellationsRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const CourseCancellationScreen(
-      initialTab: CourseNoticeTab.cancellations,
-    );
-  }
-}
-
-final class CourseNoticeMakeupsRouteData extends GoRouteData
-    with $CourseNoticeMakeupsRouteData {
-  const CourseNoticeMakeupsRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const CourseCancellationScreen(initialTab: CourseNoticeTab.makeups);
-  }
-}
-
-final class CourseNoticeRoomChangesRouteData extends GoRouteData
-    with $CourseNoticeRoomChangesRouteData {
-  const CourseNoticeRoomChangesRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const CourseCancellationScreen(
-      initialTab: CourseNoticeTab.roomChanges,
-    );
-  }
-}
-
-final class CourseSubjectsRouteData extends GoRouteData
-    with $CourseSubjectsRouteData {
-  const CourseSubjectsRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return SearchSubjectScreen(
-      onSubjectSelected: (subjectId) => unawaited(
-        CourseSubjectSyllabusRouteData(id: subjectId).push<void>(context),
-      ),
-    );
-  }
-}
-
-/// 科目詳細のパスは常にタブ付きのパスへ寄せる。
-final class CourseSubjectDetailRouteData extends GoRouteData
-    with $CourseSubjectDetailRouteData {
-  const CourseSubjectDetailRouteData({required this.id});
-
-  final String id;
-
-  @override
-  FutureOr<String?> redirect(BuildContext context, GoRouterState state) {
-    return CourseSubjectSyllabusRouteData(id: id).location;
-  }
-}
-
-final class CourseSubjectSyllabusRouteData extends GoRouteData
-    with $CourseSubjectSyllabusRouteData {
-  const CourseSubjectSyllabusRouteData({required this.id});
-
-  final String id;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return _courseSubjectDetail(
-      context,
-      id: id,
-      tab: SubjectDetailTab.syllabus,
-    );
-  }
-}
-
-final class CourseSubjectReviewsRouteData extends GoRouteData
-    with $CourseSubjectReviewsRouteData {
-  const CourseSubjectReviewsRouteData({required this.id});
-
-  final String id;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return _courseSubjectDetail(context, id: id, tab: SubjectDetailTab.reviews);
-  }
-}
-
-final class CourseSubjectReviewNewRouteData extends GoRouteData
-    with $CourseSubjectReviewNewRouteData {
-  const CourseSubjectReviewNewRouteData({required this.id});
-
-  final String id;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return SubjectReviewNewScreen(id: id);
-  }
-}
-
-final class CourseSubjectPastExamsRouteData extends GoRouteData
-    with $CourseSubjectPastExamsRouteData {
-  const CourseSubjectPastExamsRouteData({required this.id});
-
-  final String id;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return _courseSubjectDetail(
-      context,
-      id: id,
-      tab: SubjectDetailTab.pastExams,
-    );
-  }
-}
-
-final class CourseSubjectPastExamRouteData extends GoRouteData
-    with $CourseSubjectPastExamRouteData {
-  const CourseSubjectPastExamRouteData({
-    required this.id,
-    required this.pastExamId,
-  });
-
-  final String id;
-
-  /// 過去問PDFのオブジェクトキー。
-  final String pastExamId;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return CloudflarePdfViewer(
-      url: pastExamId,
-      filename: pastExamFileName(pastExamId),
-    );
-  }
-}
-
-final class CourseCalendarRouteData extends GoRouteData
-    with $CourseCalendarRouteData {
-  const CourseCalendarRouteData({required this.year});
-
-  final int year;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return CourseDocumentScreen(
-      document: CourseDocument.officialCalendar,
-      year: year,
-    );
-  }
-}
-
-final class CourseSpringTimetableRouteData extends GoRouteData
-    with $CourseSpringTimetableRouteData {
-  const CourseSpringTimetableRouteData({required this.year});
-
-  final int year;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return CourseDocumentScreen(
-      document: CourseDocument.springTimetable,
-      year: year,
-    );
-  }
-}
-
-final class CourseFallTimetableRouteData extends GoRouteData
-    with $CourseFallTimetableRouteData {
-  const CourseFallTimetableRouteData({required this.year});
-
-  final int year;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return CourseDocumentScreen(
-      document: CourseDocument.fallTimetable,
-      year: year,
-    );
-  }
-}
-
-final class FunchRouteData extends GoRouteData with $FunchRouteData {
-  const FunchRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const FunchScreen();
-  }
-}
-
-final class MapRouteData extends GoRouteData with $MapRouteData {
-  const MapRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return MapScreen(
-      onGoToSettingButtonTapped: () => const SettingsRouteData().go(context),
-    );
-  }
-}
-
-final class MapRoomRouteData extends GoRouteData with $MapRoomRouteData {
-  const MapRoomRouteData({required this.id});
-
-  final String id;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return MapScreen(
-      focusedRoomId: id,
-      onGoToSettingButtonTapped: () => const SettingsRouteData().go(context),
-    );
-  }
-}
-
-final class BusRouteData extends GoRouteData with $BusRouteData {
-  const BusRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const BusScreen();
-  }
-}
-
-final class BusStopSelectRouteData extends GoRouteData
-    with $BusStopSelectRouteData {
-  const BusStopSelectRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const BusStopSelectScreen();
-  }
-}
-
-final class BusTripRouteData extends GoRouteData with $BusTripRouteData {
-  const BusTripRouteData({required this.id});
-
-  final String id;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return BusTripScreen(id: id);
-  }
-}
-
-final class SettingsRouteData extends GoRouteData with $SettingsRouteData {
-  const SettingsRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const SettingsScreen();
-  }
-}
-
-final class AnnouncementsRouteData extends GoRouteData
-    with $AnnouncementsRouteData {
-  const AnnouncementsRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const AnnouncementScreen();
-  }
-}
-
-final class AnnouncementDetailRouteData extends GoRouteData
-    with $AnnouncementDetailRouteData {
-  const AnnouncementDetailRouteData({required this.id});
-
-  final String id;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return AnnouncementDetailScreen(id: id);
-  }
-}
-
-final class DevelopersRouteData extends GoRouteData with $DevelopersRouteData {
-  const DevelopersRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const GitHubContributorScreen();
-  }
-}
-
-final class SettingOnboardingRouteData extends GoRouteData
-    with $SettingOnboardingRouteData {
-  const SettingOnboardingRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return OnboardingScreen(onDismissed: () => context.pop());
-  }
-}
-
-final class SettingsLicenseRouteData extends GoRouteData
-    with $SettingsLicenseRouteData {
-  const SettingsLicenseRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const SettingsLicenseScreen();
-  }
-}
-
-final class DebugRouteData extends GoRouteData with $DebugRouteData {
-  const DebugRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const DebugScreen();
-  }
-}
-
-final class SubjectsRouteData extends GoRouteData with $SubjectsRouteData {
-  const SubjectsRouteData();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return SearchSubjectScreen(
-      onSubjectSelected: (subjectId) => unawaited(
-        SubjectSyllabusRouteData(id: subjectId).push<void>(context),
-      ),
-    );
-  }
-}
-
-/// 科目詳細のパスは常にタブ付きのパスへ寄せる。
-final class SubjectDetailRouteData extends GoRouteData
-    with $SubjectDetailRouteData {
-  const SubjectDetailRouteData({required this.id});
-
-  final String id;
-
-  @override
-  FutureOr<String?> redirect(BuildContext context, GoRouterState state) {
-    return SubjectSyllabusRouteData(id: id).location;
-  }
-}
-
-final class SubjectSyllabusRouteData extends GoRouteData
-    with $SubjectSyllabusRouteData {
-  const SubjectSyllabusRouteData({required this.id});
-
-  final String id;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return _subjectDetail(context, id: id, tab: SubjectDetailTab.syllabus);
-  }
-}
-
-final class SubjectReviewsRouteData extends GoRouteData
-    with $SubjectReviewsRouteData {
-  const SubjectReviewsRouteData({required this.id});
-
-  final String id;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return _subjectDetail(context, id: id, tab: SubjectDetailTab.reviews);
-  }
-}
-
-final class SubjectReviewNewRouteData extends GoRouteData
-    with $SubjectReviewNewRouteData {
-  const SubjectReviewNewRouteData({required this.id});
-
-  final String id;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return SubjectReviewNewScreen(id: id);
-  }
-}
-
-final class SubjectPastExamsRouteData extends GoRouteData
-    with $SubjectPastExamsRouteData {
-  const SubjectPastExamsRouteData({required this.id});
-
-  final String id;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return _subjectDetail(context, id: id, tab: SubjectDetailTab.pastExams);
-  }
-}
-
-final class SubjectPastExamRouteData extends GoRouteData
-    with $SubjectPastExamRouteData {
-  const SubjectPastExamRouteData({required this.id, required this.pastExamId});
-
-  final String id;
-
-  /// 過去問PDFのオブジェクトキー。
-  final String pastExamId;
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return CloudflarePdfViewer(
-      url: pastExamId,
-      filename: pastExamFileName(pastExamId),
-    );
-  }
-}
-
-/// 講義タブの科目詳細を、指定したタブを開いた状態で構築する。
-Widget _courseSubjectDetail(
-  BuildContext context, {
-  required String id,
-  required SubjectDetailTab tab,
-}) {
-  return SubjectDetailScreen(
-    id: id,
-    initialTab: tab,
-    onPastExamSelected: (objectKey) => unawaited(
-      CourseSubjectPastExamRouteData(
-        id: id,
-        pastExamId: objectKey,
-      ).push<void>(context),
-    ),
-  );
-}
-
-/// 科目検索タブの科目詳細を、指定したタブを開いた状態で構築する。
-Widget _subjectDetail(
-  BuildContext context, {
-  required String id,
-  required SubjectDetailTab tab,
-}) {
-  return SubjectDetailScreen(
-    id: id,
-    initialTab: tab,
-    onPastExamSelected: (objectKey) => unawaited(
-      SubjectPastExamRouteData(
-        id: id,
-        pastExamId: objectKey,
-      ).push<void>(context),
-    ),
-  );
-}
-
+/// タブとブランチの対応。上の `branches` に並べた順番と揃える。
 int branchIndexForTab(TabItem tab) => switch (tab) {
   TabItem.course => 0,
   TabItem.funch => 1,
